Make some portability improvements identified by Christophe: A few size_t, a couple "../Include", and a whole bunch of parenthesizing "(A && B) || (C && D)", because some compilers don't believe humans know && is higher precedence than ||.
git-svn-id: https://cvs.khronos.org/svn/repos/ogl/trunk/ecosystem/public/sdk/tools/glslang@23379 e7fa87d3-cd2b-0410-9028-fcbf551c1848

+++ glslang/MachineIndependent/ParseHelper.h
@@ -63,7 +63,7 @@ public:
                   bool forwardCompatible = false, EShMessages messages = EShMsgDefault);
 
 public:
-    bool parseShaderStrings(TPpContext&, char* strings[], int strLen[], int numStrings);
+    bool parseShaderStrings(TPpContext&, char* strings[], size_t strLen[], int numStrings);
     void parserError(const char *s);     // for bison's yyerror
 
     void C_DECL error(TSourceLoc, const char *szReason, const char *szToken,
